# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * RukoLite Online Shop UMKM MVP API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ErrorResponseError = {
  code: string;
  message: string;
  details?: unknown[];
};

export interface ErrorResponse {
  error: ErrorResponseError;
}

export interface ProductImage {
  id: string;
  url: string;
  alt?: string | null;
  sortOrder: number;
}

export interface ProductListItem {
  id: string;
  name: string;
  slug: string;
  shortDescription?: string | null;
  /** Amount in IDR integer (e.g. 150000 = Rp 150.000) */
  price: number;
  compareAtPrice?: number | null;
  stockQuantity: number;
  images: ProductImage[];
}

export interface ProductDetail {
  id: string;
  name: string;
  slug: string;
  shortDescription?: string | null;
  description?: string | null;
  price: number;
  compareAtPrice?: number | null;
  stockQuantity: number;
  images: ProductImage[];
  isActive: boolean;
}

export interface ShippingMethod {
  id: string;
  code: string;
  name: string;
  description?: string | null;
  price: number;
}

export interface StorefrontResponse {
  id: string;
  name: string;
  slug: string;
  logoUrl?: string | null;
  primaryColor: string;
  announcementText?: string | null;
  currency: string;
  country: string;
  contactEmail?: string | null;
  contactPhone?: string | null;
  activePaymentMethods: string[];
  activeShippingMethods: ShippingMethod[];
}

export interface CheckoutItem {
  productId: string;
  /** @minimum 1 */
  quantity: number;
}

export interface CheckoutCustomer {
  email: string;
  phone: string;
}

export interface CheckoutShippingAddress {
  firstName: string;
  lastName: string;
  addressLine1: string;
  addressLine2?: string;
  city: string;
  province: string;
  postalCode: string;
  country: string;
}

export type CheckoutRequestPaymentMethod = typeof CheckoutRequestPaymentMethod[keyof typeof CheckoutRequestPaymentMethod];


export const CheckoutRequestPaymentMethod = {
  manual_fake_qris: 'manual_fake_qris',
} as const;

export interface CheckoutRequest {
  /** @minItems 1 */
  items: CheckoutItem[];
  customer: CheckoutCustomer;
  shippingAddress: CheckoutShippingAddress;
  shippingMethodId: string;
  paymentMethod: CheckoutRequestPaymentMethod;
  discountCode?: string;
}

export interface PaymentInstructionDetail {
  title: string;
  description: string;
  qrPayload: string;
  expiresAt: string;
}

export interface PaymentInstruction {
  method: string;
  displayName: string;
  status: string;
  amount: number;
  instruction: PaymentInstructionDetail;
}

export interface CheckoutResponse {
  orderCode: string;
  payment: PaymentInstruction;
}

export interface OrderItem {
  id: string;
  productId?: string | null;
  itemType: string;
  nameSnapshot: string;
  skuSnapshot?: string | null;
  unitPrice: number;
  compareAtPrice?: number | null;
  quantity: number;
  lineTotal: number;
}

export interface OrderAddress {
  firstName: string;
  lastName: string;
  addressLine1: string;
  addressLine2?: string | null;
  city: string;
  province: string;
  postalCode: string;
  country: string;
}

export interface OrderResponse {
  id: string;
  orderCode: string;
  customerEmail: string;
  customerPhone: string;
  status: string;
  subtotalAmount: number;
  discountAmount: number;
  shippingAmount: number;
  totalAmount: number;
  currency: string;
  paymentMethod: string;
  items: OrderItem[];
  address: OrderAddress;
  payment: PaymentInstruction;
  createdAt: string;
}

export interface PaymentConfirmationRequest {
  payerName: string;
  reference?: string;
  note?: string;
}

export interface PaymentConfirmationResponse {
  success: boolean;
  paymentStatus: string;
}

export type GetStorefront200 = {
  data: StorefrontResponse;
};

export type ListProductsParams = {
sort?: ListProductsSort;
};

export type ListProductsSort = typeof ListProductsSort[keyof typeof ListProductsSort];


export const ListProductsSort = {
  newest: 'newest',
  price_asc: 'price_asc',
  price_desc: 'price_desc',
} as const;

export type ListProducts200 = {
  data: ProductListItem[];
};

export type GetProductBySlug200 = {
  data: ProductDetail;
};

export type ListShippingMethods200 = {
  data: ShippingMethod[];
};

export type CreateCheckout201 = {
  data: CheckoutResponse;
};

export type GetOrderByCode200 = {
  data: OrderResponse;
};

export type SubmitPaymentConfirmation200 = {
  data: PaymentConfirmationResponse;
};