# lib/api-zod/src/generated/types/checkoutResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * RukoLite Online Shop UMKM MVP API
 * OpenAPI spec version: 0.1.0
 */
import type { PaymentInstruction } from './paymentInstruction';

export interface CheckoutResponse {
  orderCode: string;
  payment: PaymentInstruction;
}