# lib/api-zod/src/generated/types/orderItem.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * RukoLite Online Shop UMKM MVP API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderItem {
  id: string;
  productId?: string | null;
  itemType: string;
  nameSnapshot: string;
  skuSnapshot?: string | null;
  unitPrice: number;
  compareAtPrice?: number | null;
  quantity: number;
  lineTotal: number;
}