# lib/api-zod/src/generated/types/orderResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * RukoLite Online Shop UMKM MVP API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderAddress } from './orderAddress';
import type { OrderItem } from './orderItem';
import type { PaymentInstruction } from './paymentInstruction';

export interface OrderResponse {
  id: string;
  orderCode: string;
  customerEmail: string;
  customerPhone: string;
  status: string;
  subtotalAmount: number;
  discountAmount: number;
  shippingAmount: number;
  totalAmount: number;
  currency: string;
  paymentMethod: string;
  items: OrderItem[];
  address: OrderAddress;
  payment: PaymentInstruction;
  createdAt: Date;
}