# lib/api-zod/src/generated/types/paymentInstructionDetail.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * RukoLite Online Shop UMKM MVP API
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentInstructionDetail {
  title: string;
  description: string;
  qrPayload: string;
  expiresAt: Date;
}