# lib/api-zod/src/generated/types/productImage.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * RukoLite Online Shop UMKM MVP API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductImage {
  id: string;
  url: string;
  alt?: string | null;
  sortOrder: number;
}